Revert "x86/efi: Request desired alignment via the PE/COFF headers"
This reverts commit aeffc4928ea21aab3c7be72f00e257799b661c29.

Linn reports that Signtool complains that kernels built with
CONFIG_EFI_STUB are violating the PE/COFF specification because the
SizeOfImage field is not a multiple of SectionAlignment.

This violation was introduced as an optimisation to skip having the
kernel relocate itself during boot and instead have the firmware place
it at a correctly aligned address.

No one else has complained and I'm not aware of any firmware
implementations that refuse to boot with commit aeffc4928ea2, but it's a
legitimate bug, so revert the offending commit.

Reported-by: Linn Crosetto <linn@hp.com>
Cc: Michael Brown <mbrown@fensystems.co.uk>
Signed-off-by: Matt Fleming <matt.fleming@intel.com>